fix(engine): address PR #9 correctness and configuration bugs
- Guard New[T] against MaxBatchSize < headerSize to prevent a guaranteed
  slice-bounds panic in the background drain goroutine on the first tick
- Cap the drain loop at math.MaxUint16 iterations before casting n to
  uint16, preventing silent header count truncation on large batches
- Fix Vyukov diff computation in push/popMarshal/pop: use unsigned
  subtraction int64(seq - pos) instead of int64(seq) - int64(pos) to
  preserve wrapping semantics and avoid signed overflow
- Return explicit packed size 20 in Tick.Marshal instead of
  unsafe.Sizeof(Tick{}) which included 4 bytes of trailing struct padding
- Migrate .golangci.yml to v2 schema (linters.settings,
  linters.exclusions.rules) and add golangci-lint config verify as a CI
  preflight step; this also surfaced pre-existing fieldalignment issues in
  Config and Batcher which are fixed by reordering pointer fields first
- Remove unimplemented Vectorized Delta Encoding from the feature list and
  update the Tick Engine section to describe the actual drain path; fix
  coalescion typo

+++ ringbuf.go
@@ -52,7 +52,7 @@ func (r *ringbuf[T]) push(item T) bool {
 		pos := r.tail.Load()
 		s := &r.data[pos&r.mask]
 		seq := s.seq.Load()
-		diff := int64(seq) - int64(pos)
+		diff := int64(seq - pos)
 
 		switch {
 		case diff == 0:
@@ -85,7 +85,7 @@ func (r *ringbuf[T]) popMarshal(buf []byte) (int, bool) {
 		pos := r.head.Load()
 		s := &r.data[pos&r.mask]
 		seq := s.seq.Load()
-		diff := int64(seq) - int64(pos+1)
+		diff := int64(seq - (pos + 1))
 
 		switch {
 		case diff == 0:
@@ -106,7 +106,7 @@ func (r *ringbuf[T]) pop(item *T) bool {
 		pos := r.head.Load()
 		s := &r.data[pos&r.mask]
 		seq := s.seq.Load()
-		diff := int64(seq) - int64(pos+1)
+		diff := int64(seq - (pos + 1))
 
 		switch {
 		case diff == 0:
